changed login function to take data from form
