+++ login.html
@@ -8,7 +8,7 @@
     <title>TSV Login</title>
     <link rel="stylesheet" href="/dist/css/main.css">
     <link rel="stylesheet" href="/src/assets/fontawsome/css/all.css">
-    <script type="module" src="/src/js/login/index.mjs"></script>
+    <script type="module" src="/src/js/login/login.mjs"></script>
 </head>
 <body>
     <header class="text-bg-primary">
@@ -49,17 +49,17 @@
     <main class="container">
       <div class="m-0 m-lg-1 p-4 bg-white border border-info rounded-5 row">
         <div class="col">
-          <form class="m-0 m-lg-1 p-4 bg-white border border-info rounded-5" action="profile.html">
+          <form class="m-0 m-lg-1 p-4 bg-white border border-info rounded-5">
             <h2 class="m-4">Sign in</h2>
             <div class="m-4">
               <label for="username">Username or email</label>
-              <input id="username" class="border rounded-1 w-100" type="text"  required minlength="5">
+              <input id="email-username" class="border rounded-1 w-100" type="text"  required minlength="5">
             </div>
             <div class="m-4">
               <label for="password">Password</label>
               <input id="password" type="password" class="border rounded-1 w-100" required minlength="10">
             </div>
-            <button type="submit" class="btn btn-primary btn-md text-secondary m-4 p-4">Login</button>
+            <button type="submit" id="login" class="btn btn-primary btn-md text-secondary m-4 p-4">Login</button>
           </form>
           <div class="m-4">
             <h2>Don't have an account?</h2>
