+++ src/js/login/login.mjs
@@ -1,33 +1,20 @@
-export async function login(userToLogin) {
+import { login } from "./login-handler.mjs";
 
-    //Login endpoint
-    const loginUrl = "https://nf-api.onrender.com/api/v1/social/auth/login";
-    
-    //Data to POST
-    const postData = {
-        method: "POST",
-        headers: {
-            "Content-Type": "application/json",
-        },
-        body: JSON.stringify(userToLogin),
+const API_BASE_URL = "https://nf-api.onrender.com";
+
+const loginUser = document.querySelector("#login");
+loginUser.addEventListener("click", function(event) {
+
+    event.preventDefault();
+
+    const email = document.querySelector("#email-username");
+    const password = document.querySelector("#password");
+
+    const loginData = {
+        email: email.value,
+        password: password.value,
     };
 
-    try {
-
-        //fetch postData
-        const response = await fetch(loginUrl, postData);
-        const json = await response.json();
-        //Get accesstoken and store it
-        const token = json.accessToken;
-        localStorage.setItem("accessToken", token);
-
-    } catch (error) {
-        console.log(error);
-    } finally {
-        if (localStorage.getItem("accessToken" === null)) {
-            console.log("failed to store token")
-        } else if (localStorage.getItem("accessToken")) {
-            console.log(localStorage.getItem("accessToken"))
-        };
-    };
-};
+    console.log(loginData)
+    login(loginData)
+});
